Moved test to core
