+++ src/main/java/io/kuzzle/sdk/core/Kuzzle.java
@@ -1,4 +1,4 @@
-package io.kuzzle.sdk.test;
+package io.kuzzle.sdk.core;
 
 import android.util.Log;
