+++ src/test/java/io/kuzzle/sdk/test/KuzzleDataCollectionTest.java
@@ -1,4 +1,4 @@
-package io.kuzzle.sdk.test;
+package io.kuzzle.sdk.core;
 
 import org.json.JSONException;
 import org.json.JSONObject;
