+++ src/test/java/io/kuzzle/sdk/test/KuzzleTest.java
@@ -1,4 +1,4 @@
-package io.kuzzle.sdk.test;
+package io.kuzzle.sdk.core;
 
 import org.json.JSONArray;
 import org.json.JSONException;
